first version of ingest webpage
