+++ src/presentation/frontend/templates/base.css
@@ -1,0 +1,7 @@
+.main-navbar {
+    background-color: #ff0000;
+}
+
+.form-select {
+    width: 160px;
+}
