+++ src/presentation/frontend/templates/circuit.html
@@ -1,0 +1,2 @@
+{% extends "base.html" %}
+{% block title %}Circuit{% endblock %}
