+++ src/presentation/frontend/templates/dataset_analysis.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %}
+{% block title %}Dataset Analysis{% endblock %}
+
+{% block content %}
+    <p>Content</p>
+    <a href="https://wikipedia.org">Ejemplo</a>
+{% endblock %}
