+++ src/presentation/frontend/templates/index.html
@@ -1,0 +1,69 @@
+{% extends "base.html" %}
+{% block title %}Tool name{% endblock %}
+
+{% block content %}
+<style>
+  .form-inline .form-control {
+    width: auto; /* Ajusta el ancho según tus necesidades */
+  }
+</style>
+
+<p>Texto</p>
+<div class="row justify-content-center">
+  <div class="col-6"> <!-- primera mitad de la pantalla -->
+    <form action="/ingest" method="post">
+      <div class="form-group container text-center">
+        <div class="row align-items-center justify-content-center">
+          <div class="col-auto">
+            <label for="language_select" class="form-label">Language:</label>
+          </div>
+          <div class="col-auto">
+            <select class="form-select" id="language_select" name="language">
+              <option selected>Default</option>
+              <option value="python">Python</option>
+              <option value="openqasm">OpenQASM</option>
+            </select>
+          </div>
+        </div>
+        <br>
+        <div class="row align-items-center justify-content-center">
+          <div class="col-auto">
+            <label for="extension_select" class="form-label">Extension:</label>
+          </div>
+          <div class="col-auto">
+            <select class="form-select" id="extension_select" name="extension">
+              <option selected>Default</option>
+              <option value="cirq">Cirq</option>
+              <option value="qiskit">Qiskit</option>
+            </select>
+          </div>
+        </div>
+        <br>
+        <div class="row">
+          <div class='col-md-6'>
+              <div class="form-group">
+                  <!-- El input donde se mostrará el datepicker -->
+                  <div class='input-group date' id='datepicker'>
+                      <input type='text' class="form-control" />
+                      <span class="input-group-addon">
+                          <span class="glyphicon glyphicon-calendar"></span>
+                      </span>
+                  </div>
+              </div>
+          </div>
+      </div>
+        <br>
+        <div class="row">
+          <div class="col text-center"><button type="submit" class="btn btn-success btn-lg" name="submit_button" value="ingest">Ingest</button></div>
+        </div>
+      </div>
+    </form>
+  </div>
+  <div class="col-6"><!-- Segunda mitad de la pantalla -->
+    <div class="progress" role="progressbar">
+      <div class="progress-bar" style="width: 75%">Progress: 75%</div>
+    </div>
+  </div>
+</div>
+
+{% endblock %}
