--- templates/index.html
@@ -1,29 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Tool name{% endblock %}
-
-{% block content %}
-    <p>Texto</p>
-    <div class="progress" role="progressbar" aria-label="Basic example" aria-valuenow="75" aria-valuemin="0" aria-valuemax="100">
-        <div class="progress-bar" style="width: 75%"></div>
-    </div>
-    <div class="Language">
-        <button class="btn btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-          Language
-        </button>
-        <ul class="dropdown-menu">
-          <li><a class="dropdown-item" href="#">Python</a></li>
-          <li><a class="dropdown-item" href="#">OpenQASM</a></li>
-        </ul>
-      </div>
-      <div class="Extension">
-        <button class="btn btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-          Extension
-        </button>
-        <ul class="dropdown-menu">
-          <li><a class="dropdown-item" href="#">Qiskit</a></li>
-          <li><a class="dropdown-item" href="#">Cirq</a></li>
-        </ul>
-      </div>
-      
-    <button type="button" class="btn btn-success">Ingest</button>
-{% endblock %}
